+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
+- Fixed bug when showing messages with pending attachements that cause loading state to be not shown in some cases.
 - Fixed clearing `MessageInputView` after dismissing message to edit
 
 ### ⬆️ Improved
@@ -77,6 +78,7 @@
 ## stream-chat-android-compose
 - Added `StreamDimens` option to the `ChatTheme`, to allow for dimension customization across the app.
 - Added localization support for the components related the channel list.
+- Added the `emptySearchContent` parameter to `ChannelList` component that allows to customize the empty placeholder, when there are no channels matching the search query.
 
 ### 🐞 Fixed